Always use invocations file
